--- app/gui/components/log_handler.py
@@ -1,13 +1,0 @@
-import logging
-import tkinter as tk
-class TextBoxLogHandler(logging.Handler):
-    def __init__(self, text_widget):
-        super().__init__()
-        self.text_widget = text_widget
-
-    def emit(self, record):
-        msg = self.format(record)
-        self.text_widget.config(state=tk.NORMAL)
-        self.text_widget.insert(tk.END, msg + '\n')
-        self.text_widget.config(state=tk.DISABLED)
-        self.text_widget.see(tk.END)
