+++ emotion_analyzer.py
